Algunas mas ideas de TAC y tipos en Types.hs
