+++ src/Playit/TAC.hs
@@ -6,23 +6,11 @@
  *  Francisco Javier    12-11163
  *  Natascha Gamboa     12-11250
 -}
+{-# LANGUAGE DatatypeContexts #-}
 module Playit.TAC where
 
-import Playit.SymbolTable
+import Playit.Types
 
 
-data ThreeAddressCode = ThreeAddressCode
-  { tacOperand :: Operation,
-    tacLvalue  :: Maybe Operand,
-    tacRvalue1 :: Maybe Operand,
-    tacRvalue2 :: Maybe Operand
-  }
-
-class SymEntryCompatible where
-  getSymID :: a -> String
-
-
-data SymEntryCompatible a => Operand a = Variable a | Constant a
-
-
-data Operation = Add | Mul -- | 
+gen :: Instr -> SymTab -> ThreeAddressCode
+gen i st = ThreeAddressCode (Operation) (Lval) (Rval1) (Rval2)
